Working with high sampling rate and big average.
This adc module was reduced to just 1 adc in order to make things
cleaner and easier.
The idea here is that there is now a static struct that is only
acessible to adc itself and when it done accumulating values of the
period, it copies this values to the external acessible struct adc0.